tambah module

+++ app/Http/Controllers/BaseController.php
@@ -14,16 +14,27 @@ class BaseController extends Controller
 	protected $model;
 	protected $base;
 	protected $baseClass;
+	protected $fields;
+	protected $judulIndex;
+	protected $judulTambah;
+	protected $judulEdit;
+	protected $deskripsiIndex;
+	protected $deskripsiTambah;
+	protected $deskripsiEdit;
+
 
     public function __construct(Model $model, $base = '')
     {
     	$this->model = $model;
     	$this->base = $base;
     	$this->baseClass = (new \ReflectionClass($this))->getShortName();
+    	$fields = $this->model->getFillable();
+    	$primaryKey = $this->model->getKeyName();
+    	$this->fields = ( in_array($primaryKey, $fields) ? [] : [null => $primaryKey] ) + $this->model->getFillable();
 
         view()->share('model', $this->model);
         view()->share('baseClass', $this->baseClass);
-    	view()->share('fields', array_map('ucwords', $this->model->getFillable()));	
+    	view()->share('fields', $this->fields);	
     	view()->share('base', $this->base);	
     	view()->share('breadcrumbLevel', 3);	
     	view()->share('breadcrumb1', 'App');
@@ -36,10 +47,15 @@ class BaseController extends Controller
         return $datatables;
     }
 
+    protected function processRequest($request)
+    {
+    	return $request;
+    }
+
     public function getIndex()
     {
-    	view()->share('judul', ucwords($this->base));	
-    	view()->share('deskripsi', 'Semua Daftar '.ucwords($this->base));	
+    	view()->share('judul', ($this->judulIndex) ? $this->judulIndex : ucwords($this->base));	
+    	view()->share('deskripsi', ($this->deskripsiIndex) ? $this->deskripsiIndex : 'Semua Daftar '.ucwords($this->base));	
     	view()->share('breadcrumb3', 'Lihat Semua');
 
     	return view('partials.appIndex');
@@ -47,7 +63,7 @@ class BaseController extends Controller
 
     public function anyData()
     {
-        $datas = $this->model->select([null => 'id']+$this->model->getFillable());
+        $datas = $this->model->select([null => $this->model->getKeyName()]+$this->model->getFillable());
 
         if ($dependencies = $this->model->dependencies()) $datas = $datas->with($dependencies);
 
@@ -56,8 +72,8 @@ class BaseController extends Controller
         $result = $datatables
             ->addColumn('menu', function($data) {
                 return 
-                '<a href="'.action($this->baseClass.'@getEdit', ['id' => $data->id]).'" class="btn btn-small btn-link"><i class="fa fa-xs fa-pencil"></i> Edit</a> '.
-                Form::open(['style' => 'display: inline!important', 'method' => 'delete', 'action' => [$this->baseClass.'@deleteHapus', $data->id]]).'  <button type="submit" onClick="return confirm(\'Yakin mau menghapus?\');" class="btn btn-small btn-link"><i class="fa fa-xs fa-trash-o"></i> Delete</button></form>';
+                '<a href="'.action($this->baseClass.'@getEdit', [$data->{$this->model->getKeyName()}]).'" class="btn btn-small btn-link"><i class="fa fa-xs fa-pencil"></i> Edit</a> '.
+                Form::open(['style' => 'display: inline!important', 'method' => 'delete', 'action' => [$this->baseClass.'@deleteHapus', $data->{$this->model->getKeyName()}]]).'  <button type="submit" onClick="return confirm(\'Yakin mau menghapus?\');" class="btn btn-small btn-link"><i class="fa fa-xs fa-trash-o"></i> Delete</button></form>';
             })
             ->make();
 
@@ -74,8 +90,8 @@ class BaseController extends Controller
         $model = $this->model;
         ${$this->base} = $model;
 
-        view()->share('judul', 'Tambah '.ucwords($this->base));
-        view()->share('deskripsi', 'Untuk menambahkan data '.$this->base);
+        view()->share('judul', ($this->judulTambah) ? $this->judulTambah : 'Tambah Data '.ucwords($this->base));
+        view()->share('deskripsi', ($this->deskripsiTambah) ? $this->deskripsiTambah : 'Untuk menambahkan data '.$this->base);
         view()->share('breadcrumb3', 'Tambah' );
         view()->share('action', 'postTambah');
 
@@ -91,6 +107,8 @@ class BaseController extends Controller
      */
     public function postTambah(Request $request)
     {
+    	$request = $this->processRequest($request);
+
         $this->validate($request, $this->model->rules());
 
         $created = $this->model->create($request->all());
@@ -109,11 +127,13 @@ class BaseController extends Controller
      */
     public function getEdit($id)
     {
-        $model = $this->model->findOrFail($id);
+        $model = $this->model->where($this->model->getKeyName(), $id)->first();
+        if ($model == null) abort('404', 'Data not found');
+        // $model = $this->model->findOrFail($id);
         ${$this->base} = $model;
 
-        view()->share('judul', 'Edit '.ucwords($this->base));
-        view()->share('deskripsi', 'Mengedit data '.$this->base);
+        view()->share('judul', ($this->judulEdit) ? $this->judulEdit : 'Edit '.ucwords($this->base));
+        view()->share('deskripsi', ($this->deskripsiEdit) ? $this->deskripsiEdit : 'Mengedit data '.$this->base);
         view()->share('breadcrumb3', 'Edit' );
         view()->share('action', 'postEdit');
         view()->share('params', compact('id'));
@@ -131,8 +151,12 @@ class BaseController extends Controller
      */
     public function postEdit(Request $request, $id)
     {
-        $model = $this->model->findOrFail($id);
+        $model = $this->model->where($this->model->getKeyName(), $id)->first();
+        if ($model == null) abort('404', 'Data not found');
+        // $model = $this->model->findOrFail($id);
 
+        $request = $this->processRequest($request);
+    	
         $this->validate($request, $model->rules());
 
         $updated = $model->update($request->all());
@@ -151,7 +175,9 @@ class BaseController extends Controller
      */
     public function deleteHapus($id)
     {
-        $model = $this->model->findOrFail($id);
+        $model = $this->model->where($this->model->getKeyName(), $id)->first();
+        if ($model == null) abort('404', 'Data not found');
+        // $model = $this->model->findOrFail($id);
 
         $deleted = $model->delete();
         return redirect($this->base);
